+++ TrainingFPTCo/Controllers/CoursesController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using TrainingFPTCo.DataDBContext;
 using TrainingFPTCo.Helpers;
+using TrainingFPTCo.Migrations;
 using TrainingFPTCo.Models;
 using TrainingFPTCo.Models.Queries;
 
@@ -10,15 +11,26 @@ namespace TrainingFPTCo.Controllers
 {
     public class CoursesController : Controller
     {
+        private readonly TrainingDbContext _dbContext;
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        private readonly string sessionRoleId;
+
+        public CoursesController(TrainingDbContext dbContext, IHttpContextAccessor httpContextAccessor)
+        {
+            _dbContext = dbContext;
+            _httpContextAccessor = httpContextAccessor;
+            sessionRoleId = _httpContextAccessor.HttpContext.Session.GetString("SessionRoleId");
+        }
+
         [HttpGet]
         public IActionResult Index()
         {
-            var sessionRoleId = HttpContext.Session.GetString("SessionRoleId");
-
             if (string.IsNullOrEmpty(HttpContext.Session.GetString("SessionUserId")))
             {
                 return RedirectToAction(nameof(LoginController.Index), "Login");
             }
+            var sessionRoleId = HttpContext.Session.GetString("SessionRoleId");
+
             if (sessionRoleId == "1" || sessionRoleId == "3" || sessionRoleId == "4")
             {
                 CourseViewModel courseModel = new CourseViewModel();
@@ -45,29 +57,43 @@ namespace TrainingFPTCo.Controllers
 
                 return View(courseModel);
             }
-
             else
             {
-                return RedirectToAction(nameof(DashboardController.Index), "Dashboard");
+                return RedirectToAction("AccessDenied", "Error");
             }
+
+
         }
         [HttpGet]
         public IActionResult Add()
         {
-            CourseDetail course = new CourseDetail();
-            List<SelectListItem> itemCategories = new List<SelectListItem>();
-            var dataCategory = new CategoryQuery().GetAllCategories(null, null);
-            foreach (var item in dataCategory)
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString("SessionUserId")))
             {
-                itemCategories.Add(new SelectListItem
-                {
-                    Value = item.Id.ToString(),
-                    Text = item.Name
-                });
+                return RedirectToAction(nameof(LoginController.Index), "Login");
             }
+            if (sessionRoleId != "4")
+            {
+                // Redirect to access denied page or return forbidden status
+                return RedirectToAction("AccessDenied", "Error");
+            }
+            else
+            {
+                List<SelectListItem> itemCategories = new List<SelectListItem>();
+                var dataCategory = new CategoryQuery().GetAllCategories(null, null);
+                foreach (var item in dataCategory)
+                {
+                    itemCategories.Add(new SelectListItem
+                    {
+                        Value = item.Id.ToString(),
+                        Text = item.Name
+                    });
+                }
+                CourseDetail course = new CourseDetail();
+                course.SessionRoleId = sessionRoleId;
 
-            ViewBag.Categories = itemCategories;
-            return View(course);
+                ViewBag.Categories = itemCategories;
+                return View(course);
+            }
         }
 
 
@@ -75,6 +101,12 @@ namespace TrainingFPTCo.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Add(CourseDetail course, IFormFile Image)
         {
+            //return Ok(course);
+            if (course.SessionRoleId != "4")
+            {
+                // Redirect to access denied page or return forbidden status
+                return RedirectToAction("AccessDenied", "Error");
+            }
             if (ModelState.IsValid)
             {
                 //xu ly insert coursr vafo database
@@ -96,6 +128,8 @@ namespace TrainingFPTCo.Controllers
                 catch (Exception ex) {
                     return Ok(ex.Message);
                 }
+                return Ok(course);
+
                 return RedirectToAction(nameof(CoursesController.Index), "Courses");
 
             }
@@ -112,11 +146,17 @@ namespace TrainingFPTCo.Controllers
             }
 
             ViewBag.Categories = itemCategories;
+
             return View(course);
         }
         [HttpPost]
         public JsonResult Delete(int id = 0)
         {
+            if (sessionRoleId != "4")
+            {
+                // Return forbidden status
+                return Json(new { cod = 403, message = "Forbidden" });
+            }
             bool deleteCourse = new CourseQuery().DeleteCourseById(id);
             if (deleteCourse)
             {
@@ -129,6 +169,10 @@ namespace TrainingFPTCo.Controllers
         [HttpGet]
         public IActionResult Update(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString("SessionUserId")))
+            {
+                return RedirectToAction(nameof(LoginController.Index), "Login");
+            }
             CourseDetail detail =new CourseQuery().GetDetailCourseById(id);
             List<SelectListItem> itemCategories = new List<SelectListItem>();
             var dataCategory = new CategoryQuery().GetAllCategories(null, null);
